pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api2mcp"
version = "0.1.0"
description = "Universal API to MCP Server Converter with LangGraph Orchestration"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Manav Ghosh", email = "manavghosh@gmail.com" }
]
keywords = ["mcp", "api", "openapi", "graphql", "converter", "langgraph", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Core
    "pyyaml>=6.0.2",          # latest: 6.0.3  (was >=6.0)
    "pydantic>=2.7.4",         # latest: 2.12.5 (was >=2.0; langgraph requires >=2.7.4)
    "httpx>=0.27.0",           # latest: 0.28.1 (was >=0.25)
    "click>=8.1.0",            # latest: 8.3.1  (was >=8.0)
    "rich>=14.0.0",            # latest: 14.3.2 (was >=13.0; MAJOR version bump)
    "jinja2>=3.1.0",           # latest: 3.1.6  (was >=3.0)
    "jsonschema>=4.23.0",      # latest: 4.26.0 (was >=4.0)
    "tenacity>=9.0.0",         # latest: 9.1.4  (was >=8.0; MAJOR version bump)
    "watchfiles>=1.0.0",       # latest: 1.1.1  (hot reload file watching; F6.2)

    # Security
    "keyring>=25.0.0",         # latest: 25.7.0 (was >=24.0)
    "cryptography>=44.0.0",    # latest: 46.0.5 (was >=41.0; security updates in 42-46)
    "pyjwt>=2.11.0",           # latest: 2.11.0 (was >=2.8;  CVE-2025-45768 in <=2.10.1)

    # MCP
    "mcp>=1.9.0,<2",           # latest: 1.26.0 (was >=1.0,<2; 1.9+ has stable streamable-HTTP)

    # Orchestration (LangGraph 1.0+)
    "langgraph>=1.0.7",        # latest: 1.0.8  (was >=1.0.0; 1.0.7 aligns with prebuilt min)
    "langgraph-prebuilt>=1.0.7",  # latest: 1.0.7 (was >=1.0.0; required by langgraph 1.0.7+)
    "langgraph-checkpoint>=4.0.0",        # latest: 4.0.0  (unchanged — already at latest)
    "langgraph-checkpoint-sqlite>=3.0.3", # latest: 3.0.3  (was >=3.0.0)
    "langchain-core>=1.2.0",   # latest: 1.2.11 (was >=1.0.0)
    "langchain-anthropic>=1.3.0",  # latest: 1.3.3 (was >=1.0.0; requires langchain-core>=1.2.11)
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",           # latest: 9.0.2  (was >=7.0;  MAJOR version bump)
    "pytest-cov>=7.0.0",       # latest: 7.0.0  (was >=4.0;  MAJOR version bump)
    "pytest-asyncio>=1.0.0",   # latest: 1.3.0  (was >=0.21; MAJOR version bump — new API)
    "pytest-mock>=3.14.0",     # latest: 3.14+  (was >=3.0)
    "respx>=0.22.0",           # latest: 0.22.0 (was >=0.20)
    "mypy>=1.14.0",            # latest: 1.19.1 (was >=1.0)
    "ruff>=0.9.0",             # latest: 0.15.x (was >=0.1;  major feature generations)
    "pre-commit>=4.0.0",       # latest: 4.5.1  (was >=3.0;  MAJOR version bump)
]
docs = [
    "mkdocs>=1.6.0",                    # latest: 1.6.1  (was >=1.5)
    "mkdocs-material>=9.5.0",           # latest: 9.7.1  (was >=9.0)
    "mkdocstrings[python]>=0.27.0",     # latest: 0.27.0  auto-generated API reference
]
graphql = [
    "graphql-core>=3.2.7",     # latest stable: 3.2.7 (3.3.0a11 is still alpha)
]
openai = [
    "langchain-openai>=0.3.0",  # GPT-4o and o-series models via LangChain
]
google = [
    "langchain-google-genai>=2.0.0",  # Gemini models via LangChain
]
postgres = [
    "langgraph-checkpoint-postgres>=3.0.4",  # latest: 3.0.4 (was >=3.0.0)
    "asyncpg>=0.30.0",         # latest: 0.31.0 (was >=0.29)
    "psycopg[binary]>=3.2.0",  # psycopg2 replaced by psycopg3; langgraph-checkpoint-postgres requires psycopg>=3.2.0
    "psycopg-pool>=3.2.0",     # required by langgraph-checkpoint-postgres
]
vault = [
    "hvac>=2.0.0",             # latest: 2.4.0  (was >=1.2;  MAJOR version bump)
]
aws = [
    "boto3>=1.35.0",           # latest: 1.42.48 (was >=1.26)
]

[project.scripts]
api2mcp = "api2mcp.cli.main:cli"

[project.urls]
Homepage = "https://github.com/manavghosh/api2mcp"
Documentation = "https://manavghosh.github.io/api2mcp"
Repository = "https://github.com/manavghosh/api2mcp"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src/api2mcp --cov-report=term-missing --cov-report=html"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "orchestration: Orchestration layer tests",
]

[tool.coverage.run]
source = ["src/api2mcp"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "langgraph.*",
    "langgraph_prebuilt.*",
    "langgraph_checkpoint.*",
    "langchain_core.*",
    "langchain_anthropic.*",
    "langchain_openai.*",
    "langchain_google_genai.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "TCH", "PTH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["api2mcp"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
